pyproject.toml
[project]
name = "workout-tracker"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "db-sqlite3>=0.0.1",
    "fastapi[standard]>=0.129.0",
    "gunicorn>=25.1.0",
    "httpx>=0.28.1",
    "python-decouple>=3.8",
    "requests>=2.32.5",
    "uvicorn>=0.41.0",
]

[dependency-groups]
dev = [
    "mypy>=1.19.1",
    "pre-commit>=4.5.1",
    "pytest>=9.0.2",
    "python-decouple-stubs>=3.8.0.4",
    "ruff>=0.15.2",
]
